+++ src/Credfeto.Cache.Proxy.Server.Tests/Helpers/PingPongTests.cs
@@ -1,0 +1,14 @@
+using Credfeto.Cache.Proxy.Server.Helpers;
+using FunFair.Test.Common;
+using Xunit;
+
+namespace Credfeto.Cache.Proxy.Server.Tests.Helpers;
+
+public sealed class PingPongTests : TestBase
+{
+    [Fact]
+    public void Model_ValueEqualsPong()
+    {
+        Assert.Equal(expected: "Pong!", actual: PingPong.Model.Value);
+    }
+}